Conflicts:
	src/main/resources/templates/ej01/listaPersonas.html

+++ src/main/java/com/jorgerubira/ejerciciosspringweb/interfaces/IEjercicio02CalculadoraService.java
@@ -5,7 +5,7 @@ public interface IEjercicio02CalculadoraService {
     
     public Integer sumar(Integer a,Integer b);
     public Integer restar(Integer a,Integer b);
-    public Integer multiplicar(Integer a,Integer b);
-    public Integer dividir(Integer a,Integer b);
+    public Double multiplicar(Double a,Double b);
+    public Double dividir(Double a,Double b);
     
 }
